Fixes #4 - Removing FloatHelper and moving stuff to NumberHelper
